requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
pydantic>=2.0
pyyaml>=6.0
requests>=2.28
cryptography>=41.0
httpx>=0.24
